Added in weapon animations and set weapon beginning animation durations, weapon switching on hot-key.

+++ Project/src/main.cpp
@@ -58,7 +58,7 @@
 /// <param name="argv">c style string of the arguments passed into the application.</param>
 int main(int argc, char** argv)
 {
-	std::srand(time(nullptr)); // generate seed using time since epoch.
+	std::srand(static_cast<unsigned int>(time(nullptr))); // generate seed using time since epoch.
 #ifdef NDEBUG
 	ShowWindow(GetConsoleWindow(), SW_HIDE);
 #endif // !NDEBUG
